% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_inference.R
\name{annotate_iGraph}
\alias{annotate_iGraph}
\title{Annotate iGraph object with centrality attributes and return the annotated iGraph.}
\usage{
annotate_iGraph(iG)
}
\arguments{
\item{iG}{igraph object created from the binary consensus matrix computed by using get_ranked_consensus_binary_matrix() function.}
}
\value{
Annotated igraph object representing the binary consensus matrix computed by get_ranked_consensus_binary_matrix() function.
}
\description{
Annotate iGraph object with centrality attributes and return the annotated iGraph.
}
\examples{
\dontrun{
annotate_iGraph(iG=inferred.iGraph)
}
}
\keyword{internal}
