% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis.R
\name{calc_effect_size_rank}
\alias{calc_effect_size_rank}
\title{Calculate and rank effect sizes from a meta-dataframe}
\usage{
calc_effect_size_rank(meta_dataframe)
}
\arguments{
\item{meta_dataframe}{A dataframe with genes on the rows (as rownames) and samples in the columns. The columns should contain (adjusted) p-values deriving from gene-based statistical testing.}
}
\value{
A gene list ranked on the base of the effect size.
}
\description{
This function calculates effect sizes based on p-values in a meta-dataframe,
and ranks the genes by their effect sizes.
}
\examples{

\dontrun{
# Example Usage:
ranked_effect_sizes <- calc_effect_size_rank(meta_dataframe)
}

}
