% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis.R
\name{calc_pvalue_based_rank}
\alias{calc_pvalue_based_rank}
\title{Calculate p-value-based rank using Fisher-based method from a meta-dataframe for meta-analysis}
\usage{
calc_pvalue_based_rank(meta_dataframe)
}
\arguments{
\item{meta_dataframe}{A dataframe with genes on the rows (as rownames) and samples in the columns. The columns should contain (adjusted) p-values deriving from gene-based statistical testing.}
}
\value{
Ranked genes based on p-value based rank.
}
\description{
This function calculates p-value-based rank using a Fisher-based method using meta-dataframe as an input,
and ranks the genes based on this p-value rank.
}
\examples{
\dontrun{
# Example Usage:
ranked_pvalues <- calc_pvalue_based_rank(meta_dataframe)
}

}
