% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis.R
\name{calc_rank_base_rank}
\alias{calc_rank_base_rank}
\title{Calculate and the genes using Rank Product method from a meta-dataframe}
\usage{
calc_rank_base_rank(meta_dataframe, class, origin)
}
\arguments{
\item{meta_dataframe}{dataframe with genes on the rows (as rownames) and samples in the columns. The columns should contain (adjusted) p-values deriving from gene-based statistical testing.}

\item{class}{A numeric vector indicating the class of each sample.  In the two class unpaired case, the label of a sample is either 0 (e.g., control group) or 1 (e.g., case group). For one class data, the label for each sample should be 1.}

\item{origin}{A character vector containing containing the origin labels of the samples.}
}
\value{
A ranked gene list using the Rank Product method.
}
\description{
This function calculates gene ranks using the Rank Product method from a meta-dataframe.
}
\examples{
\dontrun{
# Example Usage:
ranked_pvalues <- calc_rank_base_rank(meta_dataframe, class, origin)
}

}
