% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis.R
\name{compute_gsea}
\alias{compute_gsea}
\title{Computes a GSEA of a gene rank against gene sets (provided through a GMT file).}
\usage{
compute_gsea(gene_list, gmt_file, no_permutations = 10000)
}
\arguments{
\item{gene_list}{A ranked gene list.}

\item{gmt_file}{a vector containing the class labels of the samples. In the two class unpaired case, the label of a sample is either 0 (e.g., control group) or 1 (e.g., case group). For one class data, the label for each sample should be 1.}
}
\description{
Computes a GSEA of a gene rank against gene sets (provided through a GMT file).
}
\examples{
\dontrun {
compute_gsea(gene_list, gmt_file = "c2.cp.reactome.v7.2.symbols.gmt")
}
}
