% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis.R
\name{compute_gsea_thresh}
\alias{compute_gsea_thresh}
\title{Computes a threshold of the gene rank based on the enrichment score of the GSEA.}
\usage{
compute_gsea_thresh(geneList, fgsea_res, background)
}
\arguments{
\item{fgsea_res}{fgsea object deriving from fgsea results or the compute_gsea function.}

\item{background}{gmt_file or Whole gene set object deriving from the fgsea::gmtPathways(gmt_file) function.}

\item{gene_list}{result from the meta_analysis genelist or ranked genelist}
}
\description{
Computes a threshold of the gene rank based on the enrichment score of the GSEA.
}
\examples{
\dontrun {
compute_gsea_thresh(gene_list, fgsea_res, background)
}
}
