% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/late_integration.R
\name{compute_view_contributions}
\alias{compute_view_contributions}
\title{Compute Contributions of Each View to Aggregated Communities}
\usage{
compute_view_contributions(W)
}
\arguments{
\item{W}{Matrix representing the contributions of the original community memberships.}
}
\value{
A matrix representing the contributions of each view to each aggregated community.
}
\description{
This function calculates the contributions of each view (original datasets) to each of the aggregated communities.
}
