% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_annotation.R
\name{get_bubbleplot_from_pathways}
\alias{get_bubbleplot_from_pathways}
\title{Generate bubble plot of enriched pathways from gene modules.}
\usage{
get_bubbleplot_from_pathways(modules, geneID = "SYMBOL")
}
\arguments{
\item{modules}{An object containing gene modules. Should be either igraph object of class 'communities' (i.e. from get_modules function)
or an integer vector (i.e. from get_most_probable_community function) representing modules.}

\item{geneID}{The type of gene ID used in input. Default is "SYMBOL".}
}
\value{
A plot showing enriched pathways.
}
\description{
This function generates a bubble plot of enriched pathways using \code{clusterProfiler::compareCluster}
and \code{clusterProfiler::dotplot} for gene modules provided as input. It maps gene symbols to ENTREZ IDs
using \code{clusterProfiler::bitr}, and performs pathway enrichment analysis using \code{clusterProfiler::compareCluster}.
}
\examples{
# Example usage:
get_bubbleplot_from_pathways(most_probable_communities, geneID="SYMBOL")
}
