% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/late_integration.R
\name{get_most_probable_community}
\alias{get_most_probable_community}
\title{Get the most probable community from aggregated communities.}
\usage{
get_most_probable_community(aggregated_communities)
}
\arguments{
\item{aggregated_communities}{the output of function aggregate_communities containing the aggregated membership probabilities.}
}
\value{
A named integer vector where each value represents the most probable community for a gene or node,
and the names are the gene or node identifiers.
}
\description{
This function determines the most probable community for each gene or node from an input
of aggregated community memberships. It selects the community with the highest membership
probability for each gene or node and orders the results.
}
\examples{
# Example usage:
most_probable_communities <- get_most_probable_community(aggregated_communities)
}
