% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functional_annotation.R
\name{get_reactome_from_modules}
\alias{get_reactome_from_modules}
\title{Perform pathway enrichment analysis from modules using Reactome.}
\usage{
get_reactome_from_modules(
  modules,
  geneID = "SYMBOL",
  pval_cutoff = 0.05,
  outPath,
  layout = "overall"
)
}
\arguments{
\item{modules}{An object containing gene modules. Should be either igraph object of class 'communities' (i.e. from get_modules function)
or an integer vector (i.e. from get_most_probable_community function) representing modules.}

\item{geneID}{The type of gene ID used in input. Default is "SYMBOL".}

\item{pval_cutoff}{The p-value cutoff for pathway enrichment. Default is 0.05.}

\item{outPath}{Path where the output files will be exported.}

\item{layout}{The layout type for exporting results. Either "overall" or "single". Default is "overall".}
}
\value{
A data frame containing enriched pathways and associated statistics.
}
\description{
This function performs pathway enrichment analysis using Reactome for gene modules.
It takes an input of modules (either as communities or integer vector), maps gene symbols
to ENTREZ IDs using \code{clusterProfiler::bitr}, and enriches pathways using \code{ReactomePA::enrichPathway}.
Results can be exported either as individual CSV files for each module (if \code{layout = "single"})
or combined into a single CSV file (\code{Pathway_results_overall.csv}) if \code{layout = "overall"}.
}
\examples{
# Example usage:
get_reactome_from_modules(most_probable_communities, geneID="SYMBOL", pval_cutoff=0.05,
                          outPath="/path/to/output", layout="overall")
}
