% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/late_integration.R
\name{get_stacked_membership_matrix}
\alias{get_stacked_membership_matrix}
\title{Create Stacked Membership Matrix from Community Labels List}
\usage{
get_stacked_membership_matrix(community_labels_list)
}
\arguments{
\item{community_labels_list}{A list of community labels for different views.}
}
\value{
A matrix where rows represent genes and columns represent communities across all views.
}
\description{
This function constructs a stacked membership matrix from the community labels of different views.
}
