% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/late_integration.R
\name{get_unique_genes}
\alias{get_unique_genes}
\title{Extract Unique Genes from Community Labels List}
\usage{
get_unique_genes(community_labels_list)
}
\arguments{
\item{community_labels_list}{A list of community labels for different views.}
}
\value{
A vector of unique gene names.
}
\description{
This function extracts a list of unique genes from the input list of community labels.
}
