% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis.R
\name{multi_studies_adjust}
\alias{multi_studies_adjust}
\title{Batch-adjust and visualize multi-study expression data}
\usage{
multi_studies_adjust(expr_mat, samples_label, batch_labels)
}
\arguments{
\item{expr_mat}{A dataframe with genes on the rows and samples in the columns.}

\item{samples_label}{A factor of samples labels in the same order as the samples in expr_mat columns}

\item{batch_labels}{A factor of labels indicating the batches (the studies where the samples are coming from)}
}
\value{
A batch-adjusted expression matrix of the same dimension of expr_mat
}
\description{
This function takes in an expression matrix, sample labels, and batch labels,
performs batch effect adjustment using the pamr package, and visualizes the
principal component analysis (PCA) before and after batch correction.
}
\examples{
\dontrun{
# Example Usage:
adjusted_data <- multi_studies_adjust(expr_mat=expr_mat, samples_label=samples_label, batch_labels=batch_labels)
}
}
