% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_inference.R
\name{parse_edge_rank_matrix}
\alias{parse_edge_rank_matrix}
\title{Get edge rank list and binary inference matrix from edge rank matrix computed by get_ranked_consensus_matrix().}
\usage{
parse_edge_rank_matrix(
  edge_rank_matrix,
  edge_selection_strategy = "default",
  mat_weights = "rank",
  topN = 10,
  debug_output = FALSE,
  updateProgress = NULL
)
}
\arguments{
\item{edge_rank_matrix}{A symmetrix matrix with edge ranks as weight.}

\item{debug_output}{Print help and status messages to help debug the running of the function default:FALSE.}

\item{updateProgress}{Shiny application can request for update of progress from this function default:NULL.}
}
\value{
A list containing a vector of consensus edge ranks and a binary symmetrix matrix representing the edge rank matrix.
}
\description{
parse_edge_rank_matrix parses the edge rank matrix created by using the internal function get_ranked_consensus_matrix_matrix() to get a ranked edge list and a binary matrix.
}
\examples{
\dontrun{
parse_edge_rank_matrix <- function(edge_rank_matrix, debug_output=FALSE, updateProgress=NULL)
}
}
\keyword{internal}
