% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarity_network_fusion.R
\name{snf_based_integration}
\alias{snf_based_integration}
\title{Integrate Gene Expression Matrices Using Similarity Network Fusion (SNF)}
\usage{
snf_based_integration(
  list_of_gene_expression_matrices,
  K = 20,
  sigma = 0.5,
  t = 20
)
}
\arguments{
\item{list_of_gene_expression_matrices}{A list of gene expression matrices. The matrices should have genes on the rows and samples on the columns. At least two matrices should be provided.}

\item{K}{An integer specifying the number of neighbors in the K-nearest neighbors part of the algorithm. Default is 20. For more details, see the SNFtool documentation.}

\item{sigma}{A numeric value specifying the variance for the local model. Default is 0.5. For more details, see the SNFtool documentation.}

\item{t}{An integer specifying the number of iterations for the diffusion process. Default is 20. For more details, see the SNFtool documentation.}
}
\value{
A fused affinity matrix computed by the SNF algorithm. If the input validation fails (less than two matrices or no common genes), the function returns -1.
}
\description{
This function integrates multiple gene expression matrices using the Similarity Network Fusion (SNF) algorithm.
The input matrices should have genes on the rows and samples on the columns.
}
